Split out `redundant_pattern_match`
